@@ -48,7 +48,7 @@ configs_file_path = os.path.abspath(
 export_config_help(configs_file_path)
 
 project = "Modin"
-copyright = "2018-2021, Modin"
+copyright = "2018-2022, Modin Developers."
 author = "Modin contributors"
 
 # The short X.Y version
