@@ -25,9 +25,9 @@ import warnings
 
 from modin.error_message import ErrorMessage
 from modin.core.storage_formats.pandas.utils import compute_chunksize
+from modin.core.dataframe.pandas.utils import concatenate
 from modin.config import NPartitions, ProgressBar, BenchmarkMode
 
-from pandas.api.types import union_categoricals
 import os
 
 
@@ -115,7 +115,7 @@ class PandasDataframePartitionManager(ABC):
     # END Abstract Methods
 
     @classmethod
-    def column_partitions(cls, partitions):
+    def column_partitions(cls, partitions, full_axis=True):
         """
         Get the list of `BaseDataframeAxisPartition` objects representing column-wise paritions.
 
@@ -123,6 +123,8 @@ class PandasDataframePartitionManager(ABC):
         ----------
         partitions : list-like
             List of (smaller) partitions to be combined to column-wise partitions.
+        full_axis : bool, default: True
+            Whether or not this partition contains the entire column axis.
 
         Returns
         -------
@@ -137,7 +139,9 @@ class PandasDataframePartitionManager(ABC):
         if not isinstance(partitions, list):
             partitions = [partitions]
         return [
-            cls._column_partitions_class(col) for frame in partitions for col in frame.T
+            cls._column_partitions_class(col, full_axis=full_axis)
+            for frame in partitions
+            for col in frame.T
         ]
 
     @classmethod
@@ -165,7 +169,7 @@ class PandasDataframePartitionManager(ABC):
         return [cls._row_partition_class(row) for frame in partitions for row in frame]
 
     @classmethod
-    def axis_partition(cls, partitions, axis):
+    def axis_partition(cls, partitions, axis, full_axis: bool = True):
         """
         Logically partition along given axis (columns or rows).
 
@@ -175,15 +179,25 @@ class PandasDataframePartitionManager(ABC):
             List of partitions to be combined.
         axis : {0, 1}
             0 for column partitions, 1 for row partitions.
+        full_axis : bool, default: True
+            Whether or not this partition contains the entire column axis.
 
         Returns
         -------
         list
             A list of `BaseDataframeAxisPartition` objects.
         """
+        make_column_partitions = axis == 0
+        if not full_axis and not make_column_partitions:
+            raise NotImplementedError(
+                (
+                    "Row partitions must contain the entire axis. We don't "
+                    + "support virtual partitioning for row partitions yet."
+                )
+            )
         return (
             cls.column_partitions(partitions)
-            if not axis
+            if make_column_partitions
             else cls.row_partitions(partitions)
         )
 
@@ -601,32 +615,6 @@ class PandasDataframePartitionManager(ABC):
             return np.append(left_parts, right_parts, axis=axis)
 
     @classmethod
-    def concatenate(cls, dfs):
-        """
-        Concatenate pandas DataFrames with saving 'category' dtype.
-
-        Parameters
-        ----------
-        dfs : list
-            List of pandas DataFrames to concatenate.
-
-        Returns
-        -------
-        pandas.DataFrame
-            A pandas DataFrame
-        """
-        categoricals_columns = set.intersection(
-            *[set(df.select_dtypes("category").columns.tolist()) for df in dfs]
-        )
-
-        for col in categoricals_columns:
-            uc = union_categoricals([df[col] for df in dfs])
-            for df in dfs:
-                df[col] = pandas.Categorical(df[col], categories=uc.categories)
-
-        return pandas.concat(dfs)
-
-    @classmethod
     def to_pandas(cls, partitions):
         """
         Convert NumPy array of PandasDataframePartition to pandas DataFrame.
@@ -662,7 +650,7 @@ class PandasDataframePartitionManager(ABC):
         if len(df_rows) == 0:
             return pandas.DataFrame()
         else:
-            return cls.concatenate(df_rows)
+            return concatenate(df_rows)
 
     @classmethod
     def to_numpy(cls, partitions, **kwargs):
@@ -712,7 +700,8 @@ class PandasDataframePartitionManager(ABC):
 
         num_splits = NPartitions.get()
         put_func = cls._partition_class.put
-        row_chunksize, col_chunksize = compute_chunksize(df, num_splits)
+        row_chunksize = compute_chunksize(df.shape[0], num_splits)
+        col_chunksize = compute_chunksize(df.shape[1], num_splits)
 
         bar_format = (
             "{l_bar}{bar}{r_bar}"
@@ -1285,3 +1274,20 @@ class PandasDataframePartitionManager(ABC):
             Partitions of Modin Dataframe on which all deferred calls should be performed.
         """
         [part.drain_call_queue() for row in partitions for part in row]
+
+    @classmethod
+    def rebalance_partitions(cls, partitions):
+        """
+        Return the provided array of partitions without rebalancing it.
+
+        Parameters
+        ----------
+        partitions : np.ndarray
+            The 2-d array of partitions to rebalance.
+
+        Returns
+        -------
+        np.ndarray
+            The same 2-d array.
+        """
+        return partitions
