@@ -38,13 +38,22 @@ from modin.pandas.test.utils import (
     bool_arg_keys,
     bool_arg_values,
     arg_keys,
+    default_to_pandas_ignore_string,
 )
 from modin.config import NPartitions, StorageFormat
+from modin.test.test_utils import warns_that_defaulting_to_pandas
 
 NPartitions.put(4)
 
 # Force matplotlib to not use any Xwindows backend.
 matplotlib.use("Agg")
+
+# Our configuration in pytest.ini requires that we explicitly catch all
+# instances of defaulting to pandas, but some test modules, like this one,
+# have too many such instances.
+# TODO(https://github.com/modin-project/modin/issues/3655): catch all instances
+# of defaulting to pandas.
+pytestmark = pytest.mark.filterwarnings(default_to_pandas_ignore_string)
 
 
 def test_agg_dict():
@@ -94,10 +103,10 @@ def test_aggregate_alias():
 def test_aggregate_error_checking():
     modin_df = pd.DataFrame(test_data["float_nan_data"])
 
-    with pytest.warns(UserWarning):
+    with warns_that_defaulting_to_pandas():
         modin_df.aggregate({modin_df.columns[0]: "sum", modin_df.columns[1]: "mean"})
 
-    with pytest.warns(UserWarning):
+    with warns_that_defaulting_to_pandas():
         modin_df.aggregate("cumproduct")
 
     with pytest.raises(ValueError):
@@ -228,8 +237,7 @@ def test_apply_metadata():
 def test_apply_udf(data, func):
     eval_general(
         *create_test_dfs(data),
-        lambda df, *args, **kwargs: df.apply(*args, **kwargs),
-        func=func,
+        lambda df, *args, **kwargs: df.apply(func, *args, **kwargs),
         other=lambda df: df,
     )
 
