@@ -30,10 +30,19 @@ from .utils import (
     generate_multiindex,
     test_groupby_data,
     dict_equals,
+    value_equals,
+    default_to_pandas_ignore_string,
 )
 from modin.config import NPartitions
 
 NPartitions.put(4)
+
+# Our configuration in pytest.ini requires that we explicitly catch all
+# instances of defaulting to pandas, but some test modules, like this one,
+# have too many such instances.
+# TODO(https://github.com/modin-project/modin/issues/3655): catch all instances
+# of defaulting to pandas.
+pytestmark = pytest.mark.filterwarnings(default_to_pandas_ignore_string)
 
 
 def modin_groupby_equals_pandas(modin_groupby, pandas_groupby):
@@ -45,7 +54,7 @@ def modin_groupby_equals_pandas(modin_groupby, pandas_groupby):
     )
 
     for g1, g2 in itertools.zip_longest(modin_groupby, pandas_groupby):
-        assert g1[0] == g2[0]
+        value_equals(g1[0], g2[0])
         df_equals(g1[1], g2[1])
 
 
@@ -76,7 +85,7 @@ def build_types_asserter(comparator):
 
 @pytest.mark.parametrize("as_index", [True, False])
 def test_mixed_dtypes_groupby(as_index):
-    frame_data = np.random.randint(97, 198, size=(2 ** 6, 2 ** 4))
+    frame_data = np.random.randint(97, 198, size=(2**6, 2**4))
     pandas_df = pandas.DataFrame(frame_data).add_prefix("col")
     # Convert every other column to string
     for col in pandas_df.iloc[
@@ -172,7 +181,7 @@ def test_mixed_dtypes_groupby(as_index):
             {"col2": sum},
             {"col2": "max", "col4": "sum", "col5": "min"},
             {"col2": max, "col4": sum, "col5": "min"},
-            # Intersection of 'by' and agg cols for MapReduce impl
+            # Intersection of 'by' and agg cols for TreeReduce impl
             {"col0": "count", "col1": "count", "col2": "count"},
             # Intersection of 'by' and agg cols for FullAxis impl
             {"col0": "nunique", "col1": "nunique", "col2": "nunique"},
@@ -307,6 +316,9 @@ def test_simple_row_groupby(by, as_index, col1_category):
 
     if col1_category:
         pandas_df = pandas_df.astype({"col1": "category"})
+        # As of pandas 1.4.0 operators like min cause TypeErrors to be raised on unordered
+        # categorical columns. We need to specify the categorical column as ordered to bypass this.
+        pandas_df["col1"] = pandas_df["col1"].cat.as_ordered()
 
     modin_df = from_pandas(pandas_df)
     n = 1
@@ -382,7 +394,7 @@ def test_simple_row_groupby(by, as_index, col1_category):
         "max",
         min,
         sum,
-        # Intersection of 'by' and agg cols for MapReduce impl
+        # Intersection of 'by' and agg cols for TreeReduce impl
         {"col1": "count", "col2": "count"},
         # Intersection of 'by' and agg cols for FullAxis impl
         {"col1": "nunique", "col2": "nunique"},
@@ -873,13 +885,13 @@ def test_simple_col_groupby():
 
 
 @pytest.mark.parametrize(
-    "by", [np.random.randint(0, 100, size=2 ** 8), lambda x: x % 3, None]
+    "by", [np.random.randint(0, 100, size=2**8), lambda x: x % 3, None]
 )
 @pytest.mark.parametrize("as_index_series_or_dataframe", [0, 1, 2])
 def test_series_groupby(by, as_index_series_or_dataframe):
     if as_index_series_or_dataframe <= 1:
         as_index = as_index_series_or_dataframe == 1
-        series_data = np.random.randint(97, 198, size=2 ** 8)
+        series_data = np.random.randint(97, 198, size=2**8)
         modin_series = pd.Series(series_data)
         pandas_series = pandas.Series(series_data)
     else:
@@ -1280,7 +1292,7 @@ def eval_shift(modin_groupby, pandas_groupby):
 
 
 def test_groupby_on_index_values_with_loop():
-    length = 2 ** 6
+    length = 2**6
     data = {
         "a": np.random.randint(0, 100, size=length),
         "b": np.random.randint(0, 100, size=length),
@@ -1320,7 +1332,7 @@ def test_groupby_on_index_values_with_loop():
     ],
 )
 def test_groupby_multiindex(groupby_kwargs):
-    frame_data = np.random.randint(0, 100, size=(2 ** 6, 2 ** 4))
+    frame_data = np.random.randint(0, 100, size=(2**6, 2**4))
     modin_df = pd.DataFrame(frame_data)
     pandas_df = pandas.DataFrame(frame_data)
 
@@ -1348,6 +1360,75 @@ def test_groupby_multiindex(groupby_kwargs):
     # https://github.com/modin-project/modin/issues/2912
     # df_equals(md_grp.quantile(), pd_grp.quantile())
     df_equals(md_grp.first(), pd_grp.first())
+
+
+@pytest.mark.parametrize("dropna", [True, False])
+@pytest.mark.parametrize(
+    "groupby_kwargs",
+    [
+        pytest.param({"level": 1, "axis": 1}, id="level_idx_axis=1"),
+        pytest.param({"level": 1}, id="level_idx"),
+        pytest.param({"level": [1, "four"]}, id="level_idx+name"),
+        pytest.param({"by": "four"}, id="level_name"),
+        pytest.param({"by": ["one", "two"]}, id="level_name_multi_by"),
+        pytest.param(
+            {"by": ["item0", "one", "two"]},
+            id="col_name+level_name",
+        ),
+        pytest.param({"by": ["item0"]}, id="col_name"),
+        pytest.param(
+            {"by": ["item0", "item1"]},
+            id="col_name_multi_by",
+        ),
+    ],
+)
+def test_groupby_with_kwarg_dropna(groupby_kwargs, dropna):
+    modin_df = pd.DataFrame(test_data["float_nan_data"])
+    pandas_df = pandas.DataFrame(test_data["float_nan_data"])
+
+    new_index = pandas.Index([f"item{i}" for i in range(len(pandas_df))])
+    new_columns = pandas.MultiIndex.from_tuples(
+        [(i // 4, i // 2, i) for i in range(len(modin_df.columns))],
+        names=["four", "two", "one"],
+    )
+    modin_df.columns = new_columns
+    modin_df.index = new_index
+    pandas_df.columns = new_columns
+    pandas_df.index = new_index
+
+    if groupby_kwargs.get("axis", 0) == 0:
+        modin_df = modin_df.T
+        pandas_df = pandas_df.T
+
+    md_grp, pd_grp = (
+        modin_df.groupby(**groupby_kwargs, dropna=dropna),
+        pandas_df.groupby(**groupby_kwargs, dropna=dropna),
+    )
+    modin_groupby_equals_pandas(md_grp, pd_grp)
+
+    by_kwarg = groupby_kwargs.get("by", [])
+    # Disabled because of broken `dropna=False` for TreeReduce implemented aggs:
+    # https://github.com/modin-project/modin/issues/3817
+    if not (
+        not dropna
+        and len(by_kwarg) > 1
+        and any(col in modin_df.columns for col in by_kwarg)
+    ):
+        df_equals(md_grp.sum(), pd_grp.sum())
+        df_equals(md_grp.size(), pd_grp.size())
+    # Grouping on level works incorrect in case of aggregation:
+    # https://github.com/modin-project/modin/issues/2912
+    # "BaseOnPython" tests are disabled because of the bug:
+    # https://github.com/modin-project/modin/issues/3827
+    if get_current_execution() != "BaseOnPython" and any(
+        col in modin_df.columns for col in by_kwarg
+    ):
+        df_equals(md_grp.quantile(), pd_grp.quantile())
+    # Default-to-pandas tests are disabled for multi-column 'by' because of the bug:
+    # https://github.com/modin-project/modin/issues/3827
+    if not (not dropna and len(by_kwarg) > 1):
+        df_equals(md_grp.first(), pd_grp.first())
+        df_equals(md_grp._default_to_pandas(lambda df: df.sum()), pd_grp.sum())
 
 
 @pytest.mark.parametrize("groupby_axis", [0, 1])
@@ -1456,7 +1537,7 @@ def test_agg_func_None_rename(by_and_agg_dict, as_index):
 @pytest.mark.parametrize(
     "agg_fns",
     [["sum", "min", "max"], ["mean", "quantile"]],
-    ids=["reduction", "aggregation"],
+    ids=["reduce", "aggregation"],
 )
 @pytest.mark.parametrize(
     "intersection_with_by_cols",
@@ -1722,7 +1803,7 @@ def test_unknown_groupby(columns):
                     list(test_data_values[0].keys())[-1]: (sum, min, max),
                 }
             ),
-            id="Agg_and_by_intersection_MapReduce_implementation",
+            id="Agg_and_by_intersection_TreeReduce_implementation",
         ),
         pytest.param(
             lambda grp: grp.agg(
